+++ allocator.h
@@ -1,5 +1,5 @@
-#ifndef SYNTH_ALLOCATOR_H
-#define SYNTH_ALLOCATOR_H
+#ifndef CKIT_ALLOCATOR_H
+#define CKIT_ALLOCATOR_H
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -25,7 +25,7 @@
 
 typedef struct _arena {
     struct _arena* next; // Siblings
-    byte* items;
+    byte* mem;
     size_t capacity;
     size_t used;
 } Arena;
@@ -35,7 +35,7 @@ Arena* arena_new(size_t capacity) {
     Arena* a = calloc(capacity + sizeof(Arena), sizeof(byte));
     if (!a) return NULL;
 
-    a->items = ( (byte*) a) + sizeof(Arena);
+    a->mem = ( (byte*) a) + sizeof(Arena);
 
     a->capacity = capacity;
     a->used = 0;
@@ -50,8 +50,11 @@ void* arena_alloc(Arena* a, size_t alloc_size) {
 
     if (alloc_size > a->capacity) {
 
-        #ifdef SYNTH_DEBUG
-            fprintf(stderr, "[DEBUG] Requested alloc size (%zu bytes) for Arena at %p is bigger than it's capacity (%zu bytes)\n", alloc_size, (void*) a, a->capacity);
+        #ifdef CKIT_DEBUG
+        fprintf(stderr, "[DEBUG] Requested alloc size (%zu bytes) for Arena at %p is bigger "
+            "than it's capacity (%zu bytes)\n",
+            alloc_size, (void*) a, a->capacity
+        );
         #endif
 
         return NULL;
@@ -62,8 +65,11 @@ void* arena_alloc(Arena* a, size_t alloc_size) {
         if (!(a->next)) {
             a->next = arena_new(a->capacity);
 
-            #ifdef SYNTH_DEBUG
-            fprintf(stderr, "[DEBUG] Arena at address %p didn't have enough memory for alloc, requested new sibling\n", (void*) a);
+            #ifdef CKIT_DEBUG
+            fprintf(stderr, "[DEBUG] Arena at address %p didn't have enough memory for alloc, "
+                "requested new sibling\n",
+                (void*) a
+            );
             #endif
         }
 
@@ -72,7 +78,7 @@ void* arena_alloc(Arena* a, size_t alloc_size) {
         return NULL;
     }
 
-    void* ptr = (void*) ((a->items) + (a->used));
+    void* ptr = (void*) ((a->mem) + (a->used));
     a->used += alloc_size;
 
     return ptr;
@@ -82,8 +88,8 @@ bool arena_free(Arena* a) {
 
     if (!a) return false;
 
-    #ifdef SYNTH_DEBUG
-        unsigned int SYNTH_DEBUG_ARENA_SIBLING_COUNT = 0;
+    #ifdef CKIT_DEBUG
+        unsigned int CKIT_DEBUG_ARENA_SIBLING_COUNT = 0;
     #endif
 
     while (a->next) {
@@ -91,17 +97,161 @@ bool arena_free(Arena* a) {
         a->next = a->next->next;
         free(n);
 
-        #ifdef SYNTH_DEBUG
-            SYNTH_DEBUG_ARENA_SIBLING_COUNT++;
+        #ifdef CKIT_DEBUG
+            CKIT_DEBUG_ARENA_SIBLING_COUNT++;
         #endif
 
     }
 
-    #ifdef SYNTH_DEBUG
-        fprintf(stderr, "[DEBUG] Arena at address %p free'd with %u sibling(s)\n", (void*) a, SYNTH_DEBUG_ARENA_SIBLING_COUNT);
+    #ifdef CKIT_DEBUG
+        fprintf(stderr, "[DEBUG] Arena at address %p free'd with %u sibling(s)\n",
+            (void*) a, CKIT_DEBUG_ARENA_SIBLING_COUNT
+        );
     #endif
     
     free(a);
+
+    return true;
+}
+
+
+
+// ====================
+//  Bump Allocator
+// ====================
+
+typedef struct __alloc_list_bump_node {
+    struct __alloc_list_bump_node* prev;
+} __Alloc_List_Bump_Node;
+
+
+typedef struct _bump_allocator {
+    size_t capacity;
+    size_t used;
+    byte* mem;
+    __Alloc_List_Bump_Node* __alloc_list;
+    struct _bump_allocator* next; // Siblings
+} Bump_Allocator;
+
+Bump_Allocator* bump_allocator_new(size_t capacity) {
+
+    Bump_Allocator* b = calloc(capacity + sizeof(Bump_Allocator), sizeof(byte));
+    if (!b) return NULL;
+
+    b->mem = ( (byte*) b) + sizeof(Bump_Allocator);
+
+    b->capacity = capacity;
+    b->used = 0;
+    b->__alloc_list = NULL;
+    b->next = NULL;
+
+    return b;
+}
+
+void* bump_alloc(Bump_Allocator* b, size_t alloc_size) {
+
+    if (!b || alloc_size < 1) return NULL;
+
+    #ifdef CKIT_DEBUG
+    if ( (alloc_size / sizeof(__Alloc_List_Bump_Node)) <= 3 )
+        fprintf(stderr, "[DEBUG] Requested alloc for Bump_Allocator creation is inefficient, "
+            "metadata required for each alloc (%zu bytes) / alloc_size (%zu bytes) = %.2lf%%\n",
+            sizeof(__Alloc_List_Bump_Node), alloc_size,
+            100 * ( CAST(double, sizeof(__Alloc_List_Bump_Node)) / CAST(double, alloc_size)) );
+    #endif
+
+    if (alloc_size + sizeof(__Alloc_List_Bump_Node) > b->capacity) {
+
+        #ifdef CKIT_DEBUG
+        fprintf(stderr, "[DEBUG] Requested alloc size + metadata (%zu bytes) for Bump_Allocator "
+            "at %p is bigger than it's capacity (%zu bytes)\n",
+            alloc_size, (void*) b, b->capacity + sizeof(__Alloc_List_Bump_Node));
+        #endif
+
+        return NULL;
+    }
+
+    if (b->used + alloc_size + sizeof(__Alloc_List_Bump_Node) > b->capacity) {
+
+        if (!(b->next)) {
+            b->next = bump_allocator_new(b->capacity);
+
+            #ifdef CKIT_DEBUG
+            fprintf(stderr, "[DEBUG] Bump_Allocator at address %p didn't have enough memory for alloc, "
+                "requested new sibling\n", (void*) b);
+            #endif
+
+        }
+
+        if (b->next) return bump_alloc(b->next, alloc_size);
+
+        return NULL;
+    }
+
+    __Alloc_List_Bump_Node* ptr = (__Alloc_List_Bump_Node*) CAST( void*, (b->mem) + (b->used) );
+    ptr->prev = b->__alloc_list;
+    b->__alloc_list = ptr;
+
+    b->used += alloc_size + sizeof(__Alloc_List_Bump_Node);
+
+    return CAST( void*, CAST(byte*, (ptr)) + sizeof(__Alloc_List_Bump_Node) );
+}
+
+bool bump_dealloc(Bump_Allocator* b) {
+
+    if (!b) return false;
+
+    if (b->next && b->next->__alloc_list) {
+
+        #ifdef CKIT_DEBUG
+        fprintf(stderr, "[DEBUG] Bump_Allocator at address %p doesn't have the last alloc'd ptr in "
+            "its memory, checked siblings for dealloc\n",
+            (void*) b
+        );
+        #endif
+
+        return bump_dealloc(b->next);
+    }
+
+    if (!(b->__alloc_list)) return false;
+
+    __Alloc_List_Bump_Node* ptr = b->__alloc_list;
+    b->__alloc_list = b->__alloc_list->prev;
+
+    size_t size_to_dealloc = CAST( size_t, ((b->mem) + (b->used)) - (CAST(byte*, ptr)) );
+    b->used -= size_to_dealloc;
+
+    memset( CAST( void*, (b->mem) + (b->used) ), 0, size_to_dealloc);
+
+    return true;
+}
+
+bool bump_free(Bump_Allocator* b) {
+
+    if (!b) return false;
+
+    #ifdef CKIT_DEBUG
+    unsigned int CKIT_DEBUG_BUMP_SIBLING_COUNT = 0;
+    #endif
+
+    while (b->next) {
+        Bump_Allocator* n = b->next;
+        b->next = b->next->next;
+        free(n);
+
+        #ifdef CKIT_DEBUG
+        CKIT_DEBUG_BUMP_SIBLING_COUNT++;
+        #endif
+
+    }
+
+    #ifdef CKIT_DEBUG
+    fprintf(stderr, "[DEBUG] Bump_Allocator at address %p free'd with %u sibling(s)\n",
+        (void*) b, CKIT_DEBUG_BUMP_SIBLING_COUNT
+    );
+    #endif
+
+    free(b);
 
     return true;
 }
@@ -125,10 +275,19 @@ typedef struct _pool {
 } Pool;
 
 Pool* pool_new(unsigned int num_items, size_t item_size) {
-    
-    size_t actual_size = item_size + sizeof(__Free_List_Pool_Node);
 
-    Pool* p = calloc( (num_items * actual_size) + sizeof(Pool), sizeof(byte));
+    if (item_size < sizeof(__Free_List_Pool_Node)) {
+
+        item_size = sizeof(__Free_List_Pool_Node);
+
+        #ifdef CKIT_DEBUG
+        fprintf(stderr, "[DEBUG] Requested item_size for Pool creation is less than minimum "
+            "required for metadata, updated it's value to the minimum\n"
+        );
+        #endif
+    }
+
+    Pool* p = calloc( (num_items * item_size) + sizeof(Pool), sizeof(byte));
     if (!p) return NULL;
 
     p->num_items = num_items;
@@ -141,7 +300,7 @@ Pool* pool_new(unsigned int num_items, size_t item_size) {
     // Prepending nodes num_items times
     __Free_List_Pool_Node* temp;
     for (unsigned int i = num_items; i > 0; i--) {
-        temp = (__Free_List_Pool_Node*) ((p->chunks) + ((i-1) * actual_size));
+        temp = (__Free_List_Pool_Node*) ((p->chunks) + ((i-1) * item_size));
         temp->next = p->__free_list;
         p->__free_list = temp;
     }
@@ -158,8 +317,11 @@ void* pool_alloc(Pool* p) {
         if (!(p->next)) {
             p->next = pool_new(p->num_items, p->chunk_size);
 
-            #ifdef SYNTH_DEBUG
-            fprintf(stderr, "[DEBUG] Pool at address %p didn't have enough memory for alloc, requested new sibling\n", (void*) p);
+            #ifdef CKIT_DEBUG
+            fprintf(stderr, "[DEBUG] Pool at address %p didn't have enough memory for alloc, "
+                "requested new sibling\n",
+                (void*) p
+            );
             #endif
         }
 
@@ -172,7 +334,7 @@ void* pool_alloc(Pool* p) {
     p->__free_list = p->__free_list->next;
     ptr->next = NULL;
 
-    return CAST( void*, CAST(byte*, ptr) + sizeof(__Free_List_Pool_Node) );
+    return (void*) ptr;
 }
 
 bool pool_dealloc(Pool* p, void* ptr) {
@@ -180,16 +342,19 @@ bool pool_dealloc(Pool* p, void* ptr) {
     if (!p || !ptr) return false;
 
     // Separated this because it was about 165 chars long and very difficult to read
-    size_t actual_size = p->chunk_size + sizeof(__Free_List_Pool_Node);
-    byte* upper_bound = ( CAST(byte*, p->chunks) + (p->num_items * actual_size ) );
+    //size_t actual_size = p->chunk_size + sizeof(__Free_List_Pool_Node);
+    byte* upper_bound = ( (p->chunks) + (p->chunk_size * p->num_items) );
 
     // Making sure the pointer is valid to dealloc. If not, check next pools if any exists
-    if ( CAST(byte*, ptr) < CAST(byte*, p->chunks) || CAST(byte*, ptr) > upper_bound) {
+    if ( CAST(byte*, ptr) < p->chunks || CAST(byte*, ptr) >  upper_bound) {
 
         if (p->next) {
 
-            #ifdef SYNTH_DEBUG
-            fprintf(stderr, "[DEBUG] Pool at address %p doesn't have ptr %p in its memory chunks, checking sibling pools for dealloc\n", (void*) p, ptr);
+            #ifdef CKIT_DEBUG
+            fprintf(stderr, "[DEBUG] Pool at address %p doesn't have ptr %p in its memory chunks, "
+                "checked sibling pools for dealloc\n",
+                (void*) p, ptr
+            );
             #endif
 
             return pool_dealloc(p->next, ptr);
@@ -198,9 +363,9 @@ bool pool_dealloc(Pool* p, void* ptr) {
         return false;
     }
 
-    memset(ptr, 0, p->chunk_size);
+    __Free_List_Pool_Node* node_ptr = (__Free_List_Pool_Node*) ptr;
 
-    __Free_List_Pool_Node* node_ptr = CAST( __Free_List_Pool_Node*, CAST(byte*, ptr) - sizeof(__Free_List_Pool_Node) );
+    if (p->chunk_size > sizeof(__Free_List_Pool_Node)) memset(ptr, 0, p->chunk_size);
 
     node_ptr->next = p->__free_list;
     p->__free_list = node_ptr;
@@ -212,8 +377,8 @@ bool pool_free(Pool* p) {
 
     if (!p) return false;
 
-    #ifdef SYNTH_DEBUG
-        unsigned int SYNTH_DEBUG_POOL_SIBLING_COUNT = 0;
+    #ifdef CKIT_DEBUG
+    unsigned int CKIT_DEBUG_POOL_SIBLING_COUNT = 0;
     #endif
 
     while (p->next) {
@@ -221,14 +386,17 @@ bool pool_free(Pool* p) {
         p->next = p->next->next;
         free(n);
 
-        #ifdef SYNTH_DEBUG
-            SYNTH_DEBUG_POOL_SIBLING_COUNT++;
+        #ifdef CKIT_DEBUG
+        CKIT_DEBUG_POOL_SIBLING_COUNT++;
         #endif
 
     }
 
-    #ifdef SYNTH_DEBUG
-        fprintf(stderr, "[DEBUG] Pool at address %p free'd with %u sibling(s)\n", (void*) p, SYNTH_DEBUG_POOL_SIBLING_COUNT);
+    #ifdef CKIT_DEBUG
+    fprintf(stderr, "[DEBUG] Pool at address %p free'd with %u sibling(s)\n",
+        (void*) p,
+        CKIT_DEBUG_POOL_SIBLING_COUNT
+    );
     #endif
     
     free(p);